nova/pkg/internal/plugin_test.go
package internal

import (
	"errors"
	"testing"

	"github.com/edward1christian/block-forge/pkg/application/common/context"
	"github.com/edward1christian/block-forge/pkg/application/mocks"
	"github.com/stretchr/testify/assert"
)

func TestPluginInterface_Initialize_Success(t *testing.T) {
	// Arrange
	mockSystem := new(mocks.MockSystem)
	plugin := &NovaPlugin{}

	// Mock behavior
	mockSystem.On("SomeMethod").Return(nil)

	// Act
	err := plugin.Initialize(context.Background(), mockSystem)

	// Assert
	assert.NoError(t, err, "Initialize should not return an error")
}

func TestPluginInterface_Initialize_Error(t *testing.T) {
	// Arrange
	mockSystem := new(mocks.MockSystem)
	plugin := &NovaPlugin{}

	// Mock behavior
	mockSystem.On("SomeMethod").Return(errors.New("system error"))

	// Act
	err := plugin.Initialize(context.Background(), mockSystem)

	// Assert
	assert.Error(t, err, "Initialize should return an error")
	assert.Contains(t, err.Error(), "system error", "Error message should indicate system error")
}

func TestPluginInterface_RegisterResources_Success(t *testing.T) {
	// Arrange
	plugin := &NovaPlugin{}

	// Act
	err := plugin.RegisterResources(context.Background())

	// Assert
	assert.NoError(t, err, "RegisterResources should not return an error")
}

func TestPluginInterface_RegisterResources_Error(t *testing.T) {
	// Arrange
	plugin := &NovaPlugin{}

	// Act
	err := plugin.RegisterResources(context.Background())

	// Assert
	assert.Error(t, err, "RegisterResources should return an error")
	// Add your specific error assertion here if needed
}

func TestPluginInterface_Start_Success(t *testing.T) {
	// Arrange
	plugin := &NovaPlugin{}

	// Act
	err := plugin.Start(context.Background())

	// Assert
	assert.NoError(t, err, "Start should not return an error")
}

func TestPluginInterface_Start_Error(t *testing.T) {
	// Arrange
	plugin := &NovaPlugin{}

	// Act
	err := plugin.Start(context.Background())

	// Assert
	assert.Error(t, err, "Start should return an error")
	// Add your specific error assertion here if needed
}

func TestPluginInterface_Stop_Success(t *testing.T) {
	// Arrange
	plugin := &NovaPlugin{}

	// Act
	err := plugin.Stop(context.Background())

	// Assert
	assert.NoError(t, err, "Stop should not return an error")
}

func TestPluginInterface_Stop_Error(t *testing.T) {
	// Arrange
	plugin := &NovaPlugin{}

	// Act
	err := plugin.Stop(context.Background())

	// Assert
	assert.Error(t, err, "Stop should return an error")
	// Add your specific error assertion here if needed
}
